refactor: Split domain file
