+++ src/components/network-costs-trend-card.tsx
@@ -11,9 +11,10 @@ import React, { ReactNode } from "react";
 
 import { ButtonGroup } from "src/components/button-group";
 import CardSource from "src/components/card-source";
-import { PeerGroup, SunshineCostsAndTariffsData } from "src/domain/data";
 import { filterBySeparator, getPalette } from "src/domain/helpers";
 import { useQueryStateNetworkCostsTrendCardFilters } from "src/domain/query-states";
+import { SunshineCostsAndTariffsData } from "src/domain/sunshine";
+import { PeerGroup } from "src/domain/sunshine";
 import { getPeerGroupLabels } from "src/domain/translation";
 
 import { CardHeader } from "./detail-page/card";
