+++ src/components/operational-standards-card.tsx
@@ -2,7 +2,8 @@ import { Trans } from "@lingui/macro";
 import { Card, CardContent, CardProps, Typography } from "@mui/material";
 
 import CardSource from "src/components/card-source";
-import { PeerGroup, SunshineOperationalStandardsData } from "src/domain/data";
+import { SunshineOperationalStandardsData } from "src/domain/sunshine";
+import { PeerGroup } from "src/domain/sunshine";
 import { getLocalizedLabel, getPeerGroupLabels } from "src/domain/translation";
 
 import { CardHeader } from "./detail-page/card";
