+++ src/components/operational-standards-chart.tsx
@@ -12,7 +12,7 @@ import {
 } from "src/components/charts-generic/containers";
 import { HistogramColumns } from "src/components/charts-generic/histogram/histogram";
 import { Histogram } from "src/components/charts-generic/histogram/histogram-state";
-import type { SunshineOperationalStandardsData } from "src/domain/data";
+import type { SunshineOperationalStandardsData } from "src/domain/sunshine";
 import { DAYS, SWISS_FRANCS } from "src/domain/metrics";
 import { useFlag } from "src/utils/flags";
