+++ src/components/peer-group-card.tsx
@@ -2,7 +2,7 @@ import { Trans } from "@lingui/macro";
 import { Typography, Card, CardContent, CardProps } from "@mui/material";
 import React from "react";
 
-import { PeerGroup } from "src/domain/data";
+import { PeerGroup } from "src/domain/sunshine";
 import { getPeerGroupLabels } from "src/domain/translation";
 
 const PeerGroupCard: React.FC<
