+++ src/components/power-stability-chart.tsx
@@ -3,7 +3,7 @@ import { Box, BoxProps } from "@mui/material";
 import { max, mean } from "d3";
 import { useMemo, useState } from "react";
 
-import type { SunshinePowerStabilityData } from "src/domain/data";
+import type { SunshinePowerStabilityData } from "src/domain/sunshine";
 import { getTextWidth, useFormatCurrency } from "src/domain/helpers";
 import { MIN_PER_YEAR } from "src/domain/metrics";
 import { chartPalette, palette } from "src/themes/palette";
