+++ src/components/sunshine-selectors/base.tsx
@@ -2,7 +2,7 @@ import { Trans, t } from "@lingui/macro";
 import { Box, Typography } from "@mui/material";
 
 import { Combobox } from "src/components/combobox";
-import { SunshineIndicator } from "src/domain/data";
+import { SunshineIndicator } from "src/domain/sunshine";
 import { QueryStateSunshineSaidiSaifiTypology } from "src/domain/query-states";
 
 type SunshineSelectorsBaseProps = {
