+++ src/components/tariffs-trend-card.tsx
@@ -11,8 +11,9 @@ import React, { ReactNode } from "react";
 
 import { ButtonGroup } from "src/components/button-group";
 import CardSource from "src/components/card-source";
-import { PeerGroup, SunshineCostsAndTariffsData } from "src/domain/data";
 import { useQueryStateTariffsTrendCardFilters } from "src/domain/query-states";
+import { SunshineCostsAndTariffsData } from "src/domain/sunshine";
+import { PeerGroup } from "src/domain/sunshine";
 import { getPeerGroupLabels } from "src/domain/translation";
 
 import { CardHeader } from "./detail-page/card";
