+++ src/domain/data.ts
@@ -1,14 +1,6 @@
-import { extent, range, scaleThreshold } from "d3";
-import { useMemo } from "react";
-import * as z from "zod";
+import { range } from "d3";
 
 import buildEnv from "src/env/build";
-import {
-  NetworkCostsData,
-  StabilityData,
-  TariffsData,
-} from "src/graphql/resolver-types";
-import { chartPalette } from "src/themes/palette";
 
 export type ObservationValue = string | number | boolean | Date;
 export type GenericObservation = Record<string, ObservationValue>;
@@ -50,34 +42,6 @@ export type ComponentFieldsFragment =
   | ComponentFields_Measure_Fragment
   | ComponentFields_Attribute_Fragment;
 
-const getDomainFromMedianValue = (medianValue: number | undefined) => {
-  const m = medianValue ?? 0;
-  const domain = [m * 0.85, m * 0.95, m * 1.05, m * 1.15];
-  return domain;
-};
-
-export const useColorScale = <T>(options: {
-  observations: T[];
-  medianValue?: number | undefined;
-  accessor: (x: T) => number | undefined;
-}) => {
-  return useMemo(() => {
-    const domain =
-      "medianValue" in options
-        ? getDomainFromMedianValue(options.medianValue)
-        : (extent(options.observations, (d) => options.accessor(d)) as [
-            number,
-            number
-          ]);
-    const scale = scaleThreshold<number, string>()
-      .domain(domain)
-      .range(chartPalette.diverging.GreenToOrange);
-
-    return scale;
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [options.medianValue, options.observations, options.accessor]);
-};
-
 export type Entity = "municipality" | "operator" | "canton";
 
 if (!buildEnv.FIRST_PERIOD || !buildEnv.CURRENT_PERIOD) {
@@ -118,90 +82,7 @@ export const categories = [
   "C7",
 ] as const;
 
-export type PeerGroup = {
-  energyDensity: string;
-  settlementDensity: string;
-};
-
-export type NetworkLevel = {
-  id: "NE5" | "NE6" | "NE7";
-};
-
-export const asNetworkLevel = (id: string): NetworkLevel["id"] => {
-  if (id === "NE5" || id === "NE6" || id === "NE7") {
-    return id as NetworkLevel["id"];
-  }
-  throw new Error(`Invalid network level: ${id}`);
-};
-
 export type ElectricityCategory = (typeof categories)[number];
-
-export type SunshineCostsAndTariffsData = {
-  latestYear: string;
-  netTariffs: TariffsData;
-  energyTariffs: TariffsData;
-  networkCosts: NetworkCostsData;
-  operator: {
-    peerGroup: {
-      energyDensity: string;
-      settlementDensity: string;
-    };
-  };
-  updateDate: string;
-};
-
-export type SunshinePowerStabilityData = {
-  latestYear: string;
-  saidi: StabilityData;
-  saifi: StabilityData;
-
-  operator: {
-    peerGroup: {
-      energyDensity: string;
-      settlementDensity: string;
-    };
-  };
-  updateDate: string;
-};
-
-export type SunshineOperationalStandardsData = {
-  latestYear: string;
-  productVariety: {
-    ecoFriendlyProductsOffered: number;
-    productCombinationsOptions: boolean;
-    operatorsProductsOffered: {
-      operatorId: string;
-      ecoFriendlyProductsOffered: number;
-      year: string;
-    }[];
-  };
-  serviceQuality: {
-    notificationPeriodDays: number;
-    informingCustomersOfOutage: boolean;
-    operatorsNotificationPeriodDays: {
-      operatorId: string;
-      days: number;
-      year: string;
-    }[];
-  };
-
-  compliance: {
-    francsRule: string;
-    timelyPaperSubmission: boolean;
-    operatorsFrancsPerInvoice: {
-      operatorId: string;
-      francsPerInvoice: number;
-      year: string;
-    }[];
-  };
-  operator: {
-    peerGroup: {
-      energyDensity: string;
-      settlementDensity: string;
-    };
-  };
-  updateDate: string;
-};
 
 export const tariffCategories = [
   // EC
@@ -253,14 +134,3 @@ export const asTariffCategory = (category: string): TariffCategory => {
 };
 
 export type ValueFormatter = (value: number) => string;
-export const sunshineIndicatorSchema = z.enum([
-  "networkCosts",
-  "netTariffs",
-  "energyTariffs",
-  "saidi",
-  "saifi",
-  "serviceQuality",
-  "compliance",
-] as const);
-
-export type SunshineIndicator = z.infer<typeof sunshineIndicatorSchema>;
