+++ src/pages/sunshine/[entity]/[id]/operational-standards.tsx
@@ -27,7 +27,7 @@ import {
   PageParams,
   Props as SharedPageProps,
 } from "src/data/shared-page-props";
-import { SunshineOperationalStandardsData } from "src/domain/data";
+import { SunshineOperationalStandardsData } from "src/domain/sunshine";
 import { getLocalizedLabel } from "src/domain/translation";
 import { fetchOperationalStandards } from "src/lib/sunshine-data";
 import {
